SIP-23 Parse `Literal` as `SimpleType` under -Xexperimental
`LiteralType` is a `SingleType` with a literal for its path.
With an example: `1` is to `p.type`, as `1` is to `p`.

`lit` is the most precise type for an expression known to evaluate to `lit`.
Note that we cannot constant fold the expression to `lit`, since it may have a side effect.

`ConstantType(c)` is the type of an expression that can be constant folded to `c`.

When an expressions should not be constant folded, the corresponding `LiteralType`
is created: `ConstantType(c).deconst == LiteralType(c)`
(this is where all `LiteralType`s originate, except for unpickling).

`typedLiteral` always infers a `ConstantType`, which is subsequently
`deconst`ed or `widen`ed where necessary (see `widenIfNecessary`).
In short, only `final val`s may have `ConstantType`s, and
`widen`ing is required to ensure that a path is stable and accessible (roughly speaking).

When checking subtyping, this invariant is useful: `LiteralType(c).widen == c.tpe`.

TODO:
  - test! (subtyping, type inference, type tests, patmat, dependent method types)
  - fully predicate implementation on -Xsip:23
  - is it safe to reduce LITERAL in pickler?
  - symbols as literals?

Notes:
 - `()` is not a literal

Equality test for singleton type <:< AnyVal

test case:

```
val y: 5 = 5
def g(x: Int) = x match {
  case _: y.type => 0
}
```

Accept negated literal in infixTypeRest

e.g., val x : -1 => -1 = ???

TODO positions in parser

Conflicts:
	bincompat-forward.whitelist.conf
	src/compiler/scala/tools/nsc/backend/icode/TypeKinds.scala
	src/compiler/scala/tools/nsc/backend/jvm/BTypesFromSymbols.scala
	src/compiler/scala/tools/nsc/typechecker/Typers.scala